% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isatools.R
\name{isa_add_node}
\alias{isa_add_node}
\title{Insert a node to the isatab}
\usage{
isa_add_node(x, node, columns = NULL, after = NULL)
}
\arguments{
\item{x}{isatab object}

\item{node}{new node identifier (e.g. "Sample Name")}

\item{columns}{(optional) character vector with columns to add}

\item{after}{ID of the node after which the current node should be
inserted}
}
\description{
Insert a node to the isatab
}
