% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isatools.R
\name{isa_add_rows}
\alias{isa_add_rows}
\title{Add sample rows to an isatab}
\usage{
isa_add_rows(x, n, replicate = TRUE)
}
\arguments{
\item{x}{an isatab object}

\item{n}{number of rows to add}

\item{replicate}{If true (default), the values in the last row of the isatabs will be replicated.
Otherwise, empty rows will be added.}
}
\value{
isatab with expanded rows
}
\description{
Add sample rows to an isatab
}
