% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isatools.R
\name{isa_ID_find}
\alias{isa_ID_find}
\title{Find IDs of nodes or properties}
\usage{
isa_ID_find(x, node_pattern = NULL, value_pattern = NULL, prop_pattern = NULL)
}
\arguments{
\item{x}{object of class isatab}

\item{node_pattern}{return only nodes which match the given pattern}

\item{value_pattern}{return only nodes which match one of the values}

\item{prop_pattern}{return only nodes which match one of the properties}
}
\value{
Character vector of IDs
}
\description{
Find IDs of nodes or properties fullfilling specified criteria
}
\examples{
\dontrun{
isa_node_find(x, ".* Name", "alpha-N1-RNA1", "Parameter Value[Library Selection]")
}
}
