% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isatools.R
\name{isa_node_add}
\alias{isa_node_add}
\alias{isa_node_rm}
\title{Insert a node to the isatab}
\usage{
isa_node_add(x, node, columns = NULL, after = NULL)

isa_node_rm(x, node_id)
}
\arguments{
\item{x}{isatab object}

\item{node}{new node identifier (e.g. "Sample Name")}

\item{columns}{(optional) character vector with columns to add}

\item{after}{ID of the node after which the current node should be
inserted}

\item{node_id}{ID of the node to be removed}
}
\description{
Insert a node to the isatab
}
