% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isatools.R
\name{isa_nodes}
\alias{isa_nodes}
\title{Show nodes in an isatab}
\usage{
isa_nodes(x)
}
\arguments{
\item{x}{object of class \code{isatab}}
}
\value{
Returns a data frame (tibble) containing columns with node ID,
node identifier (name), number of properties associated with that
node and a summary of the values for that node.
}
\description{
Show nodes in an isatab
}
\seealso{
\code{\linkS4class{isatab}}
}
