% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isatools.R
\name{isa_to_text}
\alias{isa_to_text}
\title{Convert an investigation to a character vector}
\usage{
isa_to_text(x, i_file_name = "i_Investigation.txt")
}
\arguments{
\item{x}{object of class isa_i}

\item{i_file_name}{name of the investigation file}
}
\value{
A named character vector. Names are the file names, and values
are the contents of the files.
}
\description{
Convert an investigation to a character vector
}
\details{
This function converts an investigation object to a character vector.
Each element of the vector is a string containint the contents of a
single file. The file names are used as names of the vector elements.
}
