% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isatab_r6.R
\name{IsaTab-subset-methods}
\alias{IsaTab-subset-methods}
\alias{[[.IsaTab}
\alias{[}
\alias{IsaTab-subset}
\alias{[[}
\alias{[<-}
\alias{[[<-}
\alias{Subset}
\alias{ISA-Tab}
\alias{data}
\alias{by}
\alias{column}
\alias{ID}
\alias{Extract}
\alias{a}
\alias{node}
\alias{or}
\alias{property}
\alias{from}
\alias{the}
\alias{its}
\alias{ID.}
\alias{Note}
\alias{that}
\alias{in}
\alias{ISAtab}
\alias{standard,}
\alias{there}
\alias{can}
\alias{be}
\alias{multiple}
\alias{nodes}
\alias{with}
\alias{same}
\alias{name.}
\alias{The}
\alias{class}
\alias{assigns}
\alias{unique}
\alias{IDs}
\alias{to}
\alias{each}
\alias{and}
\alias{property.}
\alias{However,}
\alias{these}
\alias{are}
\alias{not}
\alias{part}
\alias{of}
\alias{standard}
\alias{only}
\alias{used}
\alias{internally}
\alias{R}
\alias{package.}
\title{Subset Methods for IsaTab Objects}
\usage{
\method{[[}{IsaTab}(x, i)
}
\arguments{
\item{x}{An IsaTab object}

\item{i}{The ID of the node or property to extract}
}
\value{
A data frame with the extracted node or property
}
\description{
This page provides an overview of the subset and subset-assignment methods for objects of class \code{IsaTab}.
}
\details{
The following methods are implemented:
\itemize{
\item \code{\link{[.IsaTab}}: Subset elements using \code{[}.
\item \code{\link{[[.IsaTab}}: Extract elements using \code{[[}.
\item \code{\link{[<-.IsaTab}}: Modify elements using \verb{[<-}.
\item \code{\link{[[<-.IsaTab}}: Modify nested elements using \verb{[[<-}.
}
}
\section{Subsetting}{

The \code{[}, \code{[[}, and \verb{[<-} methods allow for flexible subsetting and modification of \code{IsaTab} objects.
}

\seealso{
\code{\link{IsaTab}}

\code{\link{IsaTab-subset-methods}}
}
