% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isatab_r6.R
\name{IsaTab}
\alias{IsaTab}
\alias{summary.IsaTab}
\title{R6 Class Representing ISA-Tab Study or Assay data}
\usage{
\method{summary}{IsaTab}(object, ...)
}
\description{
blabla
}
\details{
bla bla
}
\seealso{
[<- for IsaTab: \code{\link{[<-.IsaTab}} to modify elements.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{isa_stru}}{A data frame with the ISA-Tab structure}

\item{\code{isa_data}}{A list with the ISA-Tab data}

\item{\code{type}}{The type of ISA-Tab object (study or assay)}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{n}}{A numeric vector with the number of samples in the ISA-Tab data}

\item{\code{dim}}{A numeric vector with the dimensions of the ISA-Tab data}

\item{\code{nodes}}{A character vector with the ISA-Tab nodes}

\item{\code{node_df}}{A data frame with the ISA-Tab nodes}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-IsaTab-new}{\code{IsaTab$new()}}
\item \href{#method-IsaTab-get}{\code{IsaTab$get()}}
\item \href{#method-IsaTab-set}{\code{IsaTab$set()}}
\item \href{#method-IsaTab-node_rm}{\code{IsaTab$node_rm()}}
\item \href{#method-IsaTab-find_prop_id}{\code{IsaTab$find_prop_id()}}
\item \href{#method-IsaTab-find_node_id}{\code{IsaTab$find_node_id()}}
\item \href{#method-IsaTab-find_id}{\code{IsaTab$find_id()}}
\item \href{#method-IsaTab-prop_rm}{\code{IsaTab$prop_rm()}}
\item \href{#method-IsaTab-as_data_frame}{\code{IsaTab$as_data_frame()}}
\item \href{#method-IsaTab-load_file}{\code{IsaTab$load_file()}}
\item \href{#method-IsaTab-summary}{\code{IsaTab$summary()}}
\item \href{#method-IsaTab-print}{\code{IsaTab$print()}}
\item \href{#method-IsaTab-clone}{\code{IsaTab$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IsaTab-new"></a>}}
\if{latex}{\out{\hypertarget{method-IsaTab-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IsaTab$new(type = "study")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{The type of ISA-Tab object (study or assay)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IsaTab-get"></a>}}
\if{latex}{\out{\hypertarget{method-IsaTab-get}{}}}
\subsection{Method \code{get()}}{
Get a value for a node or property in the ISA-Tab data
using the internal ID.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IsaTab$get(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The id of the node or property}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IsaTab-set"></a>}}
\if{latex}{\out{\hypertarget{method-IsaTab-set}{}}}
\subsection{Method \code{set()}}{
Set a value for a node or property in the ISA-Tab data
using the internal ID.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IsaTab$set(id, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The id of the node or property}

\item{\code{value}}{The value to set}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IsaTab-node_rm"></a>}}
\if{latex}{\out{\hypertarget{method-IsaTab-node_rm}{}}}
\subsection{Method \code{node_rm()}}{
Remove a node and its properties from the ISA-Tab data
using the internal ID.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IsaTab$node_rm(node_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{node_id}}{The id of the node to remove}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IsaTab-find_prop_id"></a>}}
\if{latex}{\out{\hypertarget{method-IsaTab-find_prop_id}{}}}
\subsection{Method \code{find_prop_id()}}{
Find the internal ID of a property given the node ID and property name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IsaTab$find_prop_id(node_id, prop_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{node_id}}{The id of the node}

\item{\code{prop_name}}{The name of the property}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IsaTab-find_node_id"></a>}}
\if{latex}{\out{\hypertarget{method-IsaTab-find_node_id}{}}}
\subsection{Method \code{find_node_id()}}{
Find the internal ID of a node based on its name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IsaTab$find_node_id(node_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{node_name}}{The name of the node}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IsaTab-find_id"></a>}}
\if{latex}{\out{\hypertarget{method-IsaTab-find_id}{}}}
\subsection{Method \code{find_id()}}{
Find the internal ID of a node or property. If the node name is ambiguous,
an error is thrown. Use the n parameter to select the node by its relative
position. For example, if there are 3 nodes called "Protocol REF", use
n = 2 to select the second one.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IsaTab$find_id(node_name, prop_name = NULL, n = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{node_name}}{The name of the node}

\item{\code{prop_name}}{The name of the property}

\item{\code{n}}{The position of the node to select (if there are multiple nodes with the same name)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IsaTab-prop_rm"></a>}}
\if{latex}{\out{\hypertarget{method-IsaTab-prop_rm}{}}}
\subsection{Method \code{prop_rm()}}{
Remove a property from the ISA-Tab data
using the internal ID.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IsaTab$prop_rm(prop_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{prop_id}}{The id of the property to remove}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IsaTab-as_data_frame"></a>}}
\if{latex}{\out{\hypertarget{method-IsaTab-as_data_frame}{}}}
\subsection{Method \code{as_data_frame()}}{
Convert IsaTab to a data frame.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IsaTab$as_data_frame()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A data frame with the ISA-Tab data
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IsaTab-load_file"></a>}}
\if{latex}{\out{\hypertarget{method-IsaTab-load_file}{}}}
\subsection{Method \code{load_file()}}{
Load an assay or study from a file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IsaTab$load_file(file)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{The file to load}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IsaTab-summary"></a>}}
\if{latex}{\out{\hypertarget{method-IsaTab-summary}{}}}
\subsection{Method \code{summary()}}{
Shows summary of the ISA-Tab data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IsaTab$summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IsaTab-print"></a>}}
\if{latex}{\out{\hypertarget{method-IsaTab-print}{}}}
\subsection{Method \code{print()}}{
Print the ISA-Tab object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IsaTab$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IsaTab-clone"></a>}}
\if{latex}{\out{\hypertarget{method-IsaTab-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IsaTab$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
