% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isatab_class.R
\name{n_row}
\alias{n_row}
\title{Generic replacement for \code{nrow()}}
\usage{
n_row(x)
}
\arguments{
\item{x}{an array-like object}
}
\value{
an integer of length 1 or NULL.
}
\description{
Generic replacement for \code{nrow()}
}
