% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listing.R
\name{node_list}
\alias{node_list}
\alias{node_show}
\title{Lists all the nodes in an isatab object}
\usage{
node_list(x)

node_show(x, node_id)
}
\arguments{
\item{x}{object of class isatab}

\item{node_id}{ID of a node to show}
}
\value{
a data.frame like object
}
\description{
Lists all the nodes in an isatab object
}
\details{
\code{node_list} returns a data frame with one row per node, showing the
number of properties associated with a given node and a summary of
values for that node.

\code{node_show} returns a data frame for a given \code{node_id} listing all
properties associated with that node and a summary of values for each of
the properties.
}
\examples{
file <- system.file("extdata", "s_isatab.txt", package="isaeditor")
isa_s <- read_isa(file)
node_list(isa_s)
}
