% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isatools.R
\name{read_isa}
\alias{read_isa}
\alias{write_isa}
\title{Read or write an isatab file}
\usage{
read_isa(file, type = "auto")

write_isa(x, file)
}
\arguments{
\item{file}{file name to read / write}

\item{type}{Either "auto", or "investigation", "study", "assay" (can
be abbreviated)}

\item{x}{isatab object}
}
\value{
\code{read_isa()} returns either an object of class \code{isatab} (for study / assay files) or
an object of class \code{isa_i} (for investigation files).
}
\description{
Read or write an isatab file
}
\examples{
\dontrun{
isa_i <- read_isa("i_Investigation.txt")
print(isa_i)

isa_a <- #' read_isa("a_isatab_transcriptome_profiling_nucleotide_sequencing.txt")
print(isa_a)
summary(isa_a)

}
}
\seealso{
\code{\linkS4class{isatab}}
}
