% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isatab_r6.R
\name{[.IsaTab}
\alias{[.IsaTab}
\title{Subset ISA-Tab data by node and property name}
\usage{
\method{[}{IsaTab}(x, node, property = NULL, n = NA)
}
\arguments{
\item{x}{An IsaTab object}

\item{node}{The name of the node to extract}

\item{property}{The name of the property to extract}

\item{n}{Which node to extract (if there are multiple nodes with the same name)}
}
\value{
A data frame with the extracted node or property
}
\description{
Subset ISA-Tab data by node and property name
}
\details{
Extract a node or property from the ISA-Tab data by its name. Note that in the
ISAtab standard, there can be multiple nodes with the same name.
Therefore, if there are multiple nodes with the same name, the \code{n} parameter
can be used to select the node by its position in the ISA-Tab structure.
}
\seealso{
\code{\link{IsaTab-subset-methods}}
}
