% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isatab_r6.R
\name{[[<-.IsaTab}
\alias{[[<-.IsaTab}
\title{Modify ISA-Tab data by column ID}
\usage{
\method{[[}{IsaTab}(x, i) <- value
}
\arguments{
\item{x}{An IsaTab object}

\item{i}{The ID of the node or property to modify}

\item{value}{The value to set}
}
\description{
Modify ISA-Tab data by column ID
}
