% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/investigation.R
\name{write_investigation}
\alias{write_investigation}
\title{Write the ISAtab investigation to a file}
\usage{
write_investigation(isa_i, file = NULL)
}
\arguments{
\item{isa_i}{The isa_i object to write}

\item{file}{The name of the file to write to. If missing, the
investigation is returned as a character vector.}
}
\value{
If file is missing, the investigation as a character vector. If
not, the file is written and the return value is invisible.
}
\description{
Write the ISAtab investigation to a file
}
