% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isatab_class.R
\name{isatab-class}
\alias{isatab-class}
\alias{print.isatab}
\alias{as.data.frame.isatab}
\alias{as_tibble.isatab}
\alias{summary.isatab}
\alias{[<-.isatab}
\alias{[.isatab}
\alias{[[<-.isatab}
\alias{[[.isatab}
\title{Class for assay and study objects}
\usage{
\method{print}{isatab}(x, ...)

\method{as.data.frame}{isatab}(x, ...)

\method{as_tibble}{isatab}(x, ...)

\method{summary}{isatab}(object, ...)

\method{[}{isatab}(x, node, col = NULL, n = 1) <- value

\method{[}{isatab}(x, node, col = NULL, n = 1)

\method{[[}{isatab}(x, col_id) <- value

\method{[[}{isatab}(x, col_id)
}
\arguments{
\item{x}{object of class isatab}

\item{...}{any further arguments are ignored}

\item{object}{object of class isatab}

\item{node}{node identifier (e.g. "Sample Name")}

\item{col}{column identifier (e.g. "Performer")}

\item{n}{instance of the matching node/column pairs}

\item{value}{vector with values which will be inserted into the isatab
at the specified column}

\item{col_id}{Column ID (e.g. "ID34")}
}
\description{
Class for isatab assay and study objects
}
\details{
Objects of this class are generated usually by reading a file with
\code{\link[=read_isa]{read_isa()}}.

Internally, it is a list containing as elements a data frame (tibble)
describing the structure of the isa tab (\code{isa_stru}) and a data frame
(tibble) containing the actual data.
\subsection{Accessing columns of an isa tab}{

There are two ways of accessing a column: by using the \code{[} function to
select a node identifier
(e.g. "Protocol REF") and, optionally, a parameter identifier (e.g.
"Performer"), or by using the \code{[[} function to select column IDs. The former has the disadvantage that
multiple identical node / parameter identifier combinations may exist,
and it may be necessary to specify which node is meant:\preformatted{isa_s <- read_isa("s_isatab.txt")
isa_s[ "Sample Name" ]
isa_s[ "Protocol REF", "Performer" ]
## 3rd instance of the combination Protocol REF / Performer
isa_s[ "Protocol REF", "Performer", 3 ]
isa_s[ "Protocol REF", "Performer", 3 ] <- "Rosalind Franklin"
}

The latter (column IDs) is not ambiguous, but column IDs are a trick
used by the package \code{isaeditor} and are not exported or read from an
actual ISA-tab. To view the column IDs, simply print the isatab object to the
screen or inspect the \code{isa_stru} element of the object:\preformatted{isa_s <- read_isa("s_isatab.txt")
isa_s
isa_s$isa_stru
isa_s[["ID34"]]
isa_s[["ID34"]] <- "Rosalind Franklin"
}
}
}
\seealso{
\link{read_isa}
}
