% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isatab_r6.R
\name{IsaTab}
\alias{IsaTab}
\alias{[.IsaTab}
\title{R6 Class Representing ISA-Tab Study or Assay data}
\usage{
\method{[}{IsaTab}(x, node, property = NULL, n = NA)
}
\description{
blabla
}
\details{
bla bla
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{isa_stru}}{A data frame with the ISA-Tab structure}

\item{\code{isa_data}}{A list with the ISA-Tab data}

\item{\code{type}}{The type of ISA-Tab object (study or assay)}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{n}}{A numeric vector with the number of samples in the ISA-Tab data}

\item{\code{dim}}{A numeric vector with the dimensions of the ISA-Tab data}

\item{\code{nodes}}{A character vector with the ISA-Tab nodes}

\item{\code{node_df}}{A data frame with the ISA-Tab nodes}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-IsaTab-new}{\code{IsaTab$new()}}
\item \href{#method-IsaTab-set}{\code{IsaTab$set()}}
\item \href{#method-IsaTab-node_rm}{\code{IsaTab$node_rm()}}
\item \href{#method-IsaTab-prop_rm}{\code{IsaTab$prop_rm()}}
\item \href{#method-IsaTab-as_data_frame}{\code{IsaTab$as_data_frame()}}
\item \href{#method-IsaTab-load_file}{\code{IsaTab$load_file()}}
\item \href{#method-IsaTab-summary}{\code{IsaTab$summary()}}
\item \href{#method-IsaTab-print}{\code{IsaTab$print()}}
\item \href{#method-IsaTab-clone}{\code{IsaTab$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IsaTab-new"></a>}}
\if{latex}{\out{\hypertarget{method-IsaTab-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IsaTab$new(type = "study")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{The type of ISA-Tab object (study or assay)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IsaTab-set"></a>}}
\if{latex}{\out{\hypertarget{method-IsaTab-set}{}}}
\subsection{Method \code{set()}}{
Set a value for a node or property in the ISA-Tab data
using the internal ID.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IsaTab$set(id, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The id of the node or property}

\item{\code{value}}{The value to set}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IsaTab-node_rm"></a>}}
\if{latex}{\out{\hypertarget{method-IsaTab-node_rm}{}}}
\subsection{Method \code{node_rm()}}{
Remove a node and its properties from the ISA-Tab data
using the internal ID.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IsaTab$node_rm(node_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{node_id}}{The id of the node to remove}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IsaTab-prop_rm"></a>}}
\if{latex}{\out{\hypertarget{method-IsaTab-prop_rm}{}}}
\subsection{Method \code{prop_rm()}}{
Remove a property from the ISA-Tab data
using the internal ID.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IsaTab$prop_rm(prop_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{prop_id}}{The id of the property to remove}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IsaTab-as_data_frame"></a>}}
\if{latex}{\out{\hypertarget{method-IsaTab-as_data_frame}{}}}
\subsection{Method \code{as_data_frame()}}{
Convert IsaTab to a data frame.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IsaTab$as_data_frame()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A data frame with the ISA-Tab data
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IsaTab-load_file"></a>}}
\if{latex}{\out{\hypertarget{method-IsaTab-load_file}{}}}
\subsection{Method \code{load_file()}}{
Load an assay or study from a file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IsaTab$load_file(file)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{The file to load}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IsaTab-summary"></a>}}
\if{latex}{\out{\hypertarget{method-IsaTab-summary}{}}}
\subsection{Method \code{summary()}}{
Shows summary of the ISA-Tab data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IsaTab$summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IsaTab-print"></a>}}
\if{latex}{\out{\hypertarget{method-IsaTab-print}{}}}
\subsection{Method \code{print()}}{
Print the ISA-Tab object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IsaTab$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IsaTab-clone"></a>}}
\if{latex}{\out{\hypertarget{method-IsaTab-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IsaTab$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
